Call pdftopng in python instead of subprocess
Using the poppler backed failed for me because I was running it from an un-activated venv and pdftopng was not in the PATH.

My original workaround was even shorter - I imported the convert function directly into the PopplerBackend class. But this import will fail on initialization if pdftopng is not installed and will prevent using the ghostscript backend, so I converted it to late import.